+++ src/components/common/upload-image-mimetypes.ts
@@ -18,3 +18,5 @@ export const supportedMimeTypes: string[] = [
   'image/tiff',
   'image/webp'
 ]
+
+export const acceptedMimeTypes = supportedMimeTypes.join(', ')