+++ src/components/editor-page/editor-pane/tool-bar/upload-image-button.tsx
@@ -11,14 +11,12 @@ import { useTranslation } from 'react-i18next'
 import { ForkAwesomeIcon } from '../../../common/fork-awesome/fork-awesome-icon'
 import { UploadInput } from '../../sidebar/upload-input'
 import { handleUpload } from '../upload-handler'
-import { supportedMimeTypes } from '../../../common/upload-image-mimetypes'
+import { acceptedMimeTypes } from '../../../common/upload-image-mimetypes'
 import { cypressId } from '../../../../utils/cypress-attribute'
 
 export interface UploadImageButtonProps {
   editor?: Editor
 }
-
-const acceptedMimeTypes = supportedMimeTypes.join(', ')
 
 export const UploadImageButton: React.FC<UploadImageButtonProps> = ({ editor }) => {
   const { t } = useTranslation()