+++ src/components/markdown-renderer/markdown-extension/task-list/task-list-replacer.tsx
@@ -18,10 +18,10 @@ export type TaskCheckedChangeHandler = (lineInMarkdown: number, checked: boolean
 export class TaskListReplacer extends ComponentReplacer {
   onTaskCheckedChange?: (lineInMarkdown: number, checked: boolean) => void
 
-  constructor(frontmatterLinesToSkip?: number, onTaskCheckedChange?: TaskCheckedChangeHandler) {
+  constructor(frontmatterLinesToSkip: number, onTaskCheckedChange?: TaskCheckedChangeHandler) {
     super()
     this.onTaskCheckedChange = (lineInMarkdown, checked) => {
-      if (onTaskCheckedChange === undefined || frontmatterLinesToSkip === undefined) {
+      if (onTaskCheckedChange === undefined) {
         return
       }
       onTaskCheckedChange(frontmatterLinesToSkip + lineInMarkdown, checked)