+++ src/redux/note-details/methods.ts
@@ -7,6 +7,7 @@
 import { store } from '..'
 import type { NoteDto } from '../../api/notes/types'
 import type {
+  ReplaceInMarkdownContentAction,
   SetNoteDetailsFromServerAction,
   SetNoteDocumentContentAction,
   UpdateNoteTitleByFirstHeadingAction,
@@ -49,6 +50,7 @@ export const updateNoteTitleByFirstHeading = (firstHeading?: string): void => {
 
 /**
  * Changes a checkbox state in the note document content. Triggered when a checkbox in the rendering is clicked.
+ *
  * @param lineInDocumentContent The line in the document content to change.
  * @param checked true if the checkbox is checked, false otherwise.
  */
@@ -58,4 +60,18 @@ export const setCheckboxInMarkdownContent = (lineInDocumentContent: number, chec
     checkboxChecked: checked,
     changedLine: lineInDocumentContent
   } as UpdateTaskListCheckboxAction)
+}
+
+/**
+ * Replaces a string in the markdown content in the global application state.
+ *
+ * @param replaceable The string that should be replaced
+ * @param replacement The replacement for the replaceable
+ */
+export const replaceInMarkdownContent = (replaceable: string, replacement: string): void => {
+  store.dispatch({
+    type: NoteDetailsActionType.REPLACE_IN_MARKDOWN_CONTENT,
+    placeholder: replaceable,
+    replacement
+  } as ReplaceInMarkdownContentAction)
 }